requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
PyYAML==6.0.3
requests==2.32.5
tqdm==4.67.1
urllib3==2.5.0
